move discussion builder in its own file
